+++ R/exec-memory.R
@@ -118,6 +118,13 @@ get_import_from_memory <- function(target, config) {
   }
 }
 
+missing_import <- function(x, config) {
+  if (is_encoded_path(x)) {
+    return(!file.exists(decode_path(x, config)))
+  }
+  identical(get_import_from_memory(x, config = config), NA_character_)
+}
+
 #' @title Get the environment where drake builds targets
 #' @description Call this function inside the commands in your plan
 #'   to get the environment where `drake` builds targets.
@@ -125,7 +132,7 @@ get_import_from_memory <- function(target, config) {
 #'   while [make()] is running. That way, you can limit the
 #'   amount of computer memory you use.
 #' @export
-#' @seealso [make()], [drake_plan()], [target()]
+#' @seealso [from_plan()]
 #' @return The environment where `drake` builds targets.
 #' @examples
 #' plan <- drake_plan(
@@ -152,7 +159,7 @@ drake_envir <- function() {
     }
     envir <- parent.frame(n = i)
   }
-  warning(
+  stop(
     "Could not find the environment where drake builds targets. ",
     "drake_envir() should only be called inside commands ",
     "in your workflow plan data frame.",
@@ -160,9 +167,62 @@ drake_envir <- function() {
   )
 }
 
-missing_import <- function(x, config) {
-  if (is_encoded_path(x)) {
-    return(!file.exists(decode_path(x, config)))
-  }
-  identical(get_import_from_memory(x, config = config), NA_character_)
+#' @title Get a target's plan info from inside a plan's command.
+#' @description Call this function inside the commands in your plan
+#'   to get an entry from any column in the plan. Changes to custom
+#'   columns referred to this way
+#'   (for example, `a` in `drake_plan(x = target(from_plan("a"), a = 123))`)
+#'   do not invalidate targets, so be careful. Only use `from_plan()`
+#'   to reference data that does not actually affect the output value
+#'   of the target. For example, you might use `from_plan()` to set the
+#'   number of parallel workers within a target:
+#'   `drake_plan(x = target(mclapply(..., from_plan("cores"), cores = 4))`.
+#'   Now, if you change the `cores` column of the plan, the parallelism will
+#'   change, but the target `x` will stay up to date.
+#' @export
+#' @seealso [drake_envir()]
+#' @return `plan[target, column]`, where `plan` is your workflow
+#'   plan data frame, `target` is the target being built,
+#'   and `column` is the name of the column of the plan you provide.
+#' @param column Character, name of a column in your `drake` plan.
+#' @examples
+#' plan <- drake_plan(my_target = target(from_plan("a"), a = "a_value"))
+#' plan
+#'
+#' cache <- storr::storr_environment()
+#' make(plan, cache = cache, session_info = FALSE)
+#' readd(my_target, cache = cache)
+#'
+#' # Why do we care?
+#' # Because we can customize parallel computing within targets this way.
+#' # Notice how we treat `mc.cores` for `mclapply()`
+#' plan <- drake_plan(
+#'   a = target(
+#'     mclapply(1:8, my_function, mc.cores = from_plan("cores")),
+#'     cores = 4
+#'   ),
+#'   b = target(
+#'     mclapply(1:4, my_function, mc.cores = from_plan("cores")),
+#'     cores = 2
+#'   )
+#' )
+#'
+#' plan
+#'
+#' # make() # nolint
+#'
+#' # Now, if you change the `cores` column of the plan, the parallelism will
+#' # change, but the targets will stay up to date.
+#' plan$cores <- c(1, 1)
+#' plan
+#'
+#' # make(plan) # nolint
+from_plan <- function(column) {
+  envir <- drake_envir()
+  plan <- get(drake_plan_marker, envir = envir)
+  target <- get(drake_target_marker, envir = envir)
+  plan[[column]][plan$target == target]
 }
+
+drake_plan_marker <- "._drake_plan"
+drake_target_marker <- "._drake_target"