+++ tests/testthat/test-memory.R
@@ -192,7 +192,7 @@ test_with_dir("drake_envir() depth", {
 })
 
 test_with_dir("drake_envir() in wrong context", {
-  expect_warning(
+  expect_error(
     drake_envir(),
     regexp = "should only be called inside commands"
   )